requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
fastapi
fastapi_microsoft_identity
python-dotenv
pyJWT
requests==2.28.2
bcrypt==4.0.1
jwt==1.3.1
pydantic==1.10.7
uuid==1.30
uvicorn==0.15.0
routers==0.1
